backend/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
packaging>=24.1
function_schema>=0.4.5
celery==5.4.0
python-decouple==3.8
redis==5.0.7
qdrant-client==1.10.1
openai>=1.56.1
pymysql==1.1.1
mysql-connector-python
mysqlclient>=2.2.4
SQLAlchemy==2.0.32
databases==0.9.0
llama-index-core<0.11.0,>=0.10.57
llama-index-llms-openai==0.1.31
llama-index>=0.10.56
langchain==0.2.14
langchain-community==0.2.12
cohere>=5.9.0
google-api-python-client==2.156.0
sentence-transformers==4.1.0